refactor: #59 put function into a separate file and adjust type
